requirements.txt
# LangGraph Grafana 可觀測性診斷專家 - 依賴包
# 核心框架
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0
langchain-community>=0.3.0

# MCP 支持
langchain-mcp-adapters>=0.1.0

# 模型支持
langchain-openai>=0.2.0

# 可視化支持
matplotlib>=3.5.0
pillow>=9.0.0

# 異步支持
asyncio-mqtt>=0.13.0

# 工具支持
requests>=2.28.0
httpx>=0.24.0

# 開發工具
python-dotenv>=1.0.0 python-dotenv>=1.0.0
